RehearsalAssistant: Change default format from AMR/NB to AAC.
Currently, we record the files as AAC files in MP4 containers with an `m4a`
extension, since some programs (e.g., WhatsApp) don't accept `mp4` files
that are audio-only to be attached.

The files that are recorded at this moment are recorded with 16kHz of sample
rate and 32kbps bitrate (obviously, this is probably way too much and needs
to be "calibrated").

This is just experimental all too experimental and we should support all
these formats with a user preference (a slider) allowing the user to choose
from a very low quality file (with low bitrate) to a higher quality file, at
the expense of a higher bitrate.